requirements.txt
streamlit
audio-recorder-streamlit
azure-cognitiveservices-speech
